requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-pptx==1.0.2
python-multipart==0.0.9
httpx==0.27.0
